+++ spec/controllers/user_ideas_controller_spec.rb
@@ -5,7 +5,7 @@ describe UserIdeasController do
 
   describe 'POST create' do
 
-    before(:each) do
+    before do
       @user = test_sign_in(Factory(:user))
     end
 
@@ -53,15 +53,9 @@ describe UserIdeasController do
           post :create, :user_idea => attr
         end.should_not change(Idea, :count)
       end
-      
-      it 'should not create an idea without a reminder' do
-        lambda do
-          post :create, :user_idea => @attr.merge(:content => 'content')
-        end.should_not change(Idea, :count)
-      end
 
       it 'should render the home page' do
-        post :create, :idea => @attr, :reminder => @reminder_attr
+        post :create, :user_idea => attr
         response.should render_template('pages/home')
       end
     end
@@ -79,71 +73,107 @@ describe UserIdeasController do
 
     describe 'success' do
 
-      before(:each) do
-        @privacy = Privacy::Values[:public]
-        @user = test_sign_in(Factory(:user))
-        @idea = Factory(:idea, :user => @user)
-        @reminder = Factory(:reminder, :user => @user, :idea => @idea, :created_at => 1.day.ago, :privacy => @privacy)
-        @idea_list = Factory(:idea_list, :user => @user)
-        @idea_list_ownership = Factory(:idea_list_ownership, :idea => @idea, :idea_list => @idea_list)
-      end
-      
-      describe 'the idea is shared with other users' do
-
-        before(:each) do
-          create_community_user
-          other_user = Factory(:user, :email => Factory.next(:email))
-          @other_reminder = Factory(:reminder, :user => other_user, :idea => @idea, :created_at => 1.day.ago, :privacy => @privacy)
-          delete :destroy, :id => @idea
-        end
-
-        it 'should donate the idea to the community' do
-          @idea.reload
-          @idea.user.name.should == 'community'
-        end
-
-        it 'should destroy all the reminders of the user that wants to delete the idea' do
-          Reminder.find_by_id(@reminder.id).should be_nil
-        end
-
-        it 'should destroy all the idea list ownerships of the idea from the user\'s lists' do
-          IdeaListOwnership.find_by_id(@idea_list_ownership.id).should be_nil
-        end
-
-        it 'should not destroy the reminders of other users' do
-          Reminder.find_by_id(@other_reminder.id).should_not be_nil
-        end  
-
+      before do
+        @user = test_sign_in Factory(:unique_user)
       end
 
-      describe 'the idea is not shared with other users' do
-        it 'should destroy the idea' do
+      context 'the idea corresponding to the user idea is owned by the user' do
+        context 'when the idea is private' do
+
+          before do
+            @idea = Factory :idea, :created_by => @user,
+                                   :owned_by => @user,
+                                   :privacy => Privacy::Values[:private]
+          end
+
+          it 'should destroy the idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should change(Idea, :count).by(-1)
+          end
+
+          it 'should destroy the user idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should change(UserIdea, :count).by(-1)
+          end
+        end
+        
+        context 'when the idea is public and shared with other users' do
+          before do
+            @idea = Factory :idea, :created_by => @user,
+                                   :owned_by => @user,
+                                   :privacy => Privacy::Values[:public]
+            other_user = Factory :unique_user
+          end
+
+          it 'should not destroy idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should_not change(Idea, :count).by(-1)
+          end
+
+          it 'should destroy the user idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should change(UserIdea, :count).by(-1)
+          end
+        end
+
+        context 'when the idea is public not shared with other users' do
+          before do
+            @idea = Factory :idea, :created_by => @user,
+                                   :owned_by => @user,
+                                   :privacy => Privacy::Values[:public]
+          end
+
+          it 'should destroy the idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should change(Idea, :count).by(-1)
+          end
+
+          it 'should destroy the user idea' do
+            lambda do 
+              delete :destroy, :id => @user_idea.id
+            end.should change(UserIdea, :count).by(-1)
+          end
+        end
+      end
+
+      context 'when the idea corresponding to the user idea is owned by some other user' do
+
+        before do
+          other_user = Factory :unique_user
+          @idea = Factory :idea, :created_by => other_user,
+                                 :owned_by => other_user,
+                                 :privacy => Privacy::Values[:public]
+        end
+
+        it 'should not destroy idea' do
           lambda do 
-            delete :destroy, :id => @idea
-          end.should change(Idea, :count).by(-1)
+            delete :destroy, :id => @user_idea.id
+          end.should_not change(Idea, :count).by(-1)
         end
 
-        it 'should destroy all it\'s reminders' do
-          Reminder.find_by_idea_id(@idea.id).should_not be_nil
-          delete :destroy, :id => @idea
-          Reminder.find_by_idea_id(@idea.id).should be_nil
-        end
-
-        it 'should destroy all the idea list ownerships of the idea from the user\'s lists' do
-          IdeaListOwnership.find_by_idea_id(@idea.id).should_not be_nil
-          delete :destroy, :id => @idea
-          IdeaListOwnership.find_by_idea_id(@idea.id).should be_nil
+        it 'should destroy the user idea' do
+          lambda do 
+            delete :destroy, :id => @user_idea.id
+          end.should change(UserIdea, :count).by(-1)
         end
       end
-
     end
 
     describe 'failure' do
 
-      it 'should deny access if the idea does not exist' do
-        test_sign_in(Factory(:user))
-        delete :destroy, :id => 9999
+      it 'should deny access if the user idea does not exist' do
+        test_sign_in Factory(:user)
+        delete :destroy, :id => @user.id
         response.should redirect_to(root_path)
+      end
+
+      it 'should deny access if the user does not own the user idea' do
+        pending
       end
     end
   end